pyproject.toml
[tool.poetry]
name = "hypatys"
version = "0.1.0"
description = "Data processing system for an investigation system."
authors = ["Julio Cesar"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
asyncpg = "^0.27.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
python-multipart = "^0.0.5"
sqlalchemy = "^2.0.0"
bcrypt = "^4.2.1"
alembic = "^1.14.0"
boto3 = "^1.35.97"
confluent-kafka = "^2.8.0"
aiokafka = "^0.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^3.0.0"
autoflake = "^2.0.0"
flake8 = "^6.0.0"
mypy = "^1.4.0"
isort = "^5.12.0"
pre-commit = "^3.3.0"
black = "^23.0.0"

[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
